+++ app/components/styles/globals.css
