+++ app/layout.tsx
@@ -1,8 +1,8 @@
-import './globals.css'
 import type { Metadata } from 'next'
 import { Inter, Roboto, Poppins } from 'next/font/google'
-import Navbar from './navbar'
-import Subnav from './subnav'
+import Navbar from './components/navbar'
+import Subnav from './components/subnav'
+import './components/styles/globals.css'
 
 const inter = Inter({ subsets: ['latin'] })
 const roboto = Roboto({
